Made seperate class for ContactAdapter. Improved code quality of ContactsActivity and BlockControllerUnitTest
